+++ kamu-cli/src/commands/pull_images_command.rs
@@ -2,11 +2,17 @@ use super::{Command, Error};
 use kamu::infra::utils::docker_client::DockerClient;
 use kamu::infra::utils::docker_images;
 
-pub struct PullImagesCommand {}
+pub struct PullImagesCommand {
+    container_runtime: DockerClient,
+    pull_test_deps: bool,
+}
 
 impl PullImagesCommand {
-    pub fn new<'a>() -> Self {
-        Self {}
+    pub fn new<'a>(container_runtime: DockerClient, pull_test_deps: bool) -> Self {
+        Self {
+            container_runtime,
+            pull_test_deps,
+        }
     }
 }
 
@@ -16,17 +22,23 @@ impl Command for PullImagesCommand {
     }
 
     fn run(&mut self) -> Result<(), Error> {
-        let docker_client = DockerClient::new();
-
-        let images = [
+        let mut images = vec![
             docker_images::SPARK,
             docker_images::FLINK,
             docker_images::JUPYTER,
         ];
 
+        if self.pull_test_deps {
+            images.extend(vec![
+                docker_images::HTTPD,
+                docker_images::FTP,
+                docker_images::MINIO,
+            ])
+        }
+
         for img in images.iter() {
             eprintln!("{}: {}", console::style("Pulling image").bold(), img);
-            docker_client.pull_cmd(img).status()?;
+            self.container_runtime.pull_cmd(img).status()?;
         }
 
         Ok(())