+++ kamu-core/src/infra/utils/docker_images.rs
@@ -1,9 +1,9 @@
-pub const SPARK: &str = "kamudata/engine-spark:0.10.2";
+pub const SPARK: &str = "docker.io/kamudata/engine-spark:0.10.2";
 pub const LIVY: &str = SPARK;
-pub const FLINK: &str = "kamudata/engine-flink:0.8.0-flink_1.13.1-scala_2.12-java8";
-pub const JUPYTER: &str = "kamudata/jupyter-uber:0.0.1";
+pub const FLINK: &str = "docker.io/kamudata/engine-flink:0.8.0-flink_1.13.1-scala_2.12-java8";
+pub const JUPYTER: &str = "docker.io/kamudata/jupyter-uber:0.0.1";
 
 // Test Images
-pub const HTTPD: &str = "httpd:2.4";
-pub const FTP: &str = "bogem/ftp";
-pub const MINIO: &str = "minio/minio";
+pub const HTTPD: &str = "docker.io/httpd:2.4";
+pub const FTP: &str = "docker.io/bogem/ftp";
+pub const MINIO: &str = "docker.io/minio/minio";